refactor: rename 'ActionCard' to 'ManageSoftLiquidation'
