+++ packages/curve-ui-kit/src/features/manage-soft-liquidation/index.tsx
@@ -1,3 +1,3 @@
-export { ManageSoftLiquidation, type Props } from './ui/ActionCard'
+export { ManageSoftLiquidation, type Props } from './ui/ManageSoftLiquidation'
 export type { Props as ImproveHealthProps } from './ui/tabs/ImproveHealth'
 export type { Props as ClosePositionProps } from './ui/tabs/ClosePosition'
